refactor: rename sidebarElemetTypes to sidebar
